+++ db/updates/update_12.5-12.6.sql
@@ -1,4 +1,4 @@
 ALTER TABLE `makeFormula` ADD `updated_at` TIMESTAMP on update CURRENT_TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP AFTER `created_at`; 
 ALTER TABLE `suppliers` CHANGE `updated_at_at` `updated_at` TIMESTAMP on update CURRENT_TIMESTAMP NULL DEFAULT CURRENT_TIMESTAMP; 
 ALTER TABLE `pv_meta` CHANGE `updated_at_at` `updated_at` TIMESTAMP on update CURRENT_TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP; 
-
+ALTER TABLE `user_prefs` CHANGE `owner_id` `owner_id` VARCHAR(255) NOT NULL;