+++ scripts/entrypoint.sh
@@ -61,11 +61,15 @@ while [ $RETRY_COUNT -lt $MAX_RETRIES ]; do
             echo "update_db_schema.sh script failed. Please check the logs."
         fi
 
-        # Start the session monitor
-        /usr/bin/session_monitor &
-        if [ $? -ne 0 ]; then
+        # Start the session monitor if SESSION_MONITOR is true
+        if [ "$SESSION_MONITOR" = "true" ]; then
+            /usr/bin/session_monitor &
+            if [ $? -ne 0 ]; then
             echo "Failed to start session monitor. Exiting."
             exit 1
+            fi
+        else
+            echo "Session monitor is disabled."
         fi
         
         # Execute the sync_db.sh script